Fix ctx enter completion
ctx enter completed the context name multiple times because the value of
the options variable was preserved across calls to _ctx_complete. When
the command line read `ctx enter foo` then, upon subsequent completions,
options was equal to "enter". This caused the context name to be
completed multiple times.

Fix #3.