Support terminal multiplexers
Previously tmux and screen were not supported. Starting a new session
would reset context information.

This commit makes _setup enter the context specified in CTX_NAME if the
shell is interactive. Starting a tmux or screen session works as
expected - the context of the parent shell is entered automatically.

Fix #5