New implementation under the hood + libs
This commit is a major overhaul of bash-ctx inner working. It used to
make changes to the current shell process by loading a context script
file. Now it starts a new bash process in order to keep the contexts
fully isolated.

On top of that a concept of libs has been implemented. libs are
libraries of shared functionality that can be loaded in any context
(think git aliases, etc.)